requirements/base.txt
Django
django-cleanup
Pillow
django-environ
djangorestframework
django-cors-headers
djangorestframework-simplejwt
djoser
